fix(tianmu) There will be a problem of too many tuples when querying multiple tables
The logic of this modification is as follows:

Descriptor object, MIUpdatingIterator object and DimensionVector are all newly defined objects this time. If you want iterators to work, you must have corresponding DimGroups. Because I didn't fully understand the usage of these classes, I used (dims. SetAll()) to set them, causing all DimGroups to converge.

The modification method is to use the corresponding virtual column to set DimGroups, so that the data volume is at most the data volume of the current table