Fix errors in the Q4 (exists subquery) query result
Q4 uses the EXISTS subquery
and previously optimized the EXISTS scenario using a parallel Hash Join,
but an exception was found in the result set.
Roll back the EXISTS process to a NEstd loop,
and the query result is restored.
This time in advance to ensure the correctness.
On the basis of ensuring correctness,
the optimization of performance is considered.