feat(handler): support alter column length and column type #1418
[summary]
different with MySQL 5.7, some features has deprecated.
`NO_AUTO_CREATE_USER` introduced in 5.7 and is removed in MySQL 8.0.
'utf8' is currently an alias for the character set UTF8MB3, but will be an alias
for UTF8MB4 in a future release. Please consider using UTF8MB4 in order to be unambiguous.