fix(tianmu): revert code, mv ret value from try block back to catch block
[summary]

The logic of this modification is as follows:
Previously, set ret action has been moved to try block, which is not efficient because every time we do truncate success, ret will be setted to 1. This time we will move the set ret action back to catch block, which will only trigged when truncate failed.