fix(tiannmu): To revert impl for semi-join in tianmu layer
The impl of semi-join in tianmu has bugs and it's not right way to impl that.
So that, at firstly, we revert these lines, and then will re-impl it. And
refine some codes to make it more concise and clear. And, thd is added as an
input param for some functions.