fix(tianmu): Fix up the unknown exception after instance killed randomly (#1841)
If the instance was killed by `kill -9 'pid'` randomly, the new inserted data will not be written into `DATA` file under
the param of `tianmu_insert_delayed` was set to 0. Under this configuration, the data will write to memory not a `DATA`
file. Therefore, when the instance was killed and restarted, the data will be lost, and the `DATA` can not be found, and
an TianmuError exception will be thrown.

changed the writing behavior of instance from writting data into memory to writting data into `DATA` file immediately as
that of `tianmu_insert_delayed=1`.