fix(tianmu): To fixup the instance crashed if the result of aggregate function goes out of bounds
To caculate `sum(length())` will lead a corruption in destructor of `ValueOrNull`, frees an array of
char. Before the array deletes, the validity of that should be checked, and after that the pointer of
that arrary should also be set to nullptr, which make sure it's a safe code piece.